requirements.txt
ansiblemetrics==0.3.2
gitpython==3.1.3
imbalanced-learn==0.6.2
joblib==0.15.1
matplotlib==3.2.1
nltk==3.5
numpy==1.18.5
pandas==1.0.4
pydriller==1.13
pygithub==1.51
python-dotenv==0.13.0
scikit-learn==0.23.1
repository-scorer==1.0.3
requests==2.23.0
treeinterpreter==0.2.2
pyyaml==5.3.1

